+++ code/cell.compiler/src/node.bin/util/index.ts
@@ -1,3 +1,4 @@
+export * from './util.hash';
 export * from './util.logger';
 export * from './util.loadConfig';
 export * from './util.ts';
