+++ code/cell.compiler/src/node.bin/util/util.logger.ts
@@ -1,5 +1,5 @@
-import { log, ModuleFederationPlugin, t } from '../node/common';
-import { COMMANDS } from './constants';
+import { log, ModuleFederationPlugin, t } from '../common';
+import { COMMANDS } from '../constants';
 import { loadConfig } from './util.loadConfig';
 
 type B = t.BuilderChain<t.CompilerModelMethods>;
